Added PostButton file and optimized the code
