+++ frontend/src/component/Post.jsx
@@ -1,20 +1,20 @@
 import React, { useState } from 'react';
 import styles from './Post.module.css';
+import PostButtons from './PostButtons';
 import Heart from 'react-animated-heart';
 
+
+
 export default function Post() {
-  const [isClick, setClick] = useState(false);
-  const [likeCount, setLikeCount] = useState(0);
   const [isFollowing, setIsFollowing] = useState(false);
 
 
   const handleFollowToggle = () => {
     setIsFollowing((prevState) => !prevState);
   };
-  const handleLikeToggle = () => {
-    setClick(!isClick);
-    setLikeCount(likeCount + (isClick ? -1 : 1));
-  };
+
+
+
 
   return (
     <div className={styles.post}>
@@ -30,6 +30,7 @@ export default function Post() {
         <i className="fa-regular fa-image"></i>
         <i className="fa-regular fa-face-smile"></i>
         <i className="fa-regular fa-calendar-days"></i>
+
         <button className={styles.Postbtn}>Post</button>
       </div>
       <hr />
@@ -44,17 +45,10 @@ export default function Post() {
           </div>
         </div>
         <img src="https://media.licdn.com/dms/image/D4D12AQG56UPUtPJj0w/article-cover_image-shrink_600_2000/0/1669373321238?e=2147483647&v=beta&t=uUMorY6yjVASSX_Iy-VNu8eXkAMfkmY97PWk7zDKR0w" alt="news " className={styles.pstimg} />
-        <div className={styles.reactions}>
-          <div className={styles.likeContainer}>
-            <Heart isClick={isClick} onClick={() => handleLikeToggle()} />
-            <span className={styles.likeCount}>{likeCount}</span>
-          </div>
-          <i className="fa-regular fa-comments"></i>
-          <i className="fa-regular fa-share-from-square"></i>
-        </div>
+        <PostButtons/>
       </div>
 
-   
+
       <div className={styles.userpost}>
         <div className={styles.profileContainer}>
           <div className={styles.usrprofile}>
@@ -65,14 +59,8 @@ export default function Post() {
           </div>
         </div>
         <img src="https://media.licdn.com/dms/image/D4D12AQG56UPUtPJj0w/article-cover_image-shrink_600_2000/0/1669373321238?e=2147483647&v=beta&t=uUMorY6yjVASSX_Iy-VNu8eXkAMfkmY97PWk7zDKR0w" alt="news " className={styles.pstimg} />
-        <div className={styles.reactions}>
-          <div className={styles.likeContainer}>
-            <Heart isClick={isClick} onClick={() => handleLikeToggle()} />
-            <span className={styles.likeCount}>{likeCount}</span>
-          </div>
-          <i className="fa-regular fa-comments"></i>
-          <i className="fa-regular fa-share-from-square"></i>
-        </div>
+        <PostButtons/>
+
       </div>
       <div className={styles.userpost}>
         <div className={styles.profileContainer}>
@@ -84,14 +72,8 @@ export default function Post() {
           </div>
         </div>
         <img src="https://media.licdn.com/dms/image/D4D12AQG56UPUtPJj0w/article-cover_image-shrink_600_2000/0/1669373321238?e=2147483647&v=beta&t=uUMorY6yjVASSX_Iy-VNu8eXkAMfkmY97PWk7zDKR0w" alt="news " className={styles.pstimg} />
-        <div className={styles.reactions}>
-          <div className={styles.likeContainer}>
-            <Heart isClick={isClick} onClick={() => handleLikeToggle()} />
-            <span className={styles.likeCount}>{likeCount}</span>
-          </div>
-          <i className="fa-regular fa-comments"></i>
-          <i className="fa-regular fa-share-from-square"></i>
-        </div>
+        <PostButtons/>
+
       </div>
 
       <div className={styles.userpost}>
@@ -104,14 +86,8 @@ export default function Post() {
           </div>
         </div>
         <img src="https://media.licdn.com/dms/image/D4D12AQG56UPUtPJj0w/article-cover_image-shrink_600_2000/0/1669373321238?e=2147483647&v=beta&t=uUMorY6yjVASSX_Iy-VNu8eXkAMfkmY97PWk7zDKR0w" alt="news " className={styles.pstimg} />
-        <div className={styles.reactions}>
-          <div className={styles.likeContainer}>
-            <Heart isClick={isClick} onClick={() => handleLikeToggle()} />
-            <span className={styles.likeCount}>{likeCount}</span>
-          </div>
-          <i className="fa-regular fa-comments"></i>
-          <i className="fa-regular fa-share-from-square"></i>
-        </div>
+        <PostButtons/>
+
       </div>
     </div>
   );
